feat(refactor): move update code out of github module
