+++ pkg/git/github.go
@@ -1,17 +1,9 @@
 package git
 
 import (
-	"archive/tar"
-	"compress/gzip"
 	"context"
-	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"path/filepath"
 	"regexp"
-	"runtime"
 	"strings"
 
 	"github.com/google/go-github/v59/github"
@@ -55,119 +47,6 @@ func DeriveRepositoryFullnameFromUrl(githubRepoUrl string) (RepositoryFullname, 
 		organization: orgName,
 		name:         repoName,
 	}, nil
-}
-
-type CoreCtlAsset struct {
-	Version   string
-	Url       string
-	Changelog string
-}
-
-func GetLatestCorectlAsset(release *github.RepositoryRelease) (*CoreCtlAsset, error) {
-	dummyAsset := CoreCtlAsset{}
-	if release.Assets == nil {
-		return &dummyAsset, errors.New("no assets found for the latest release")
-	}
-
-	architecture := runtime.GOARCH
-
-	// Required due to the goreleaser config
-	if architecture == "amd64" {
-		architecture = "x86_64"
-	}
-	targetAssetName := fmt.Sprintf("corectl_%s_%s.tar.gz", runtime.GOOS, architecture)
-	for _, asset := range release.Assets {
-		assetName := strings.ToLower(asset.GetName())
-		if assetName == targetAssetName {
-			log.Debug().Str("asset", assetName).Msg("github: found release asset with matching architecture & os")
-			return &CoreCtlAsset{
-				Url:       *asset.BrowserDownloadURL,
-				Version:   *release.TagName,
-				Changelog: *release.Body,
-			}, nil
-		}
-	}
-
-	return &dummyAsset, errors.New("no asset found for the current architecture and OS")
-
-}
-
-func GetLatestCorectlRelease(client *github.Client) (*github.RepositoryRelease, error) {
-	dummyRelease := github.RepositoryRelease{}
-	release, _, err := client.Repositories.GetLatestRelease(context.Background(), "coreeng", "corectl")
-	if err != nil {
-		return &dummyRelease, err
-	}
-	return release, nil
-}
-func GetCorectlReleaseByTag(client *github.Client, version string) (*github.RepositoryRelease, error) {
-	dummyRelease := github.RepositoryRelease{}
-	release, _, err := client.Repositories.GetReleaseByTag(context.Background(), "coreeng", "corectl", version)
-	if err != nil {
-		return &dummyRelease, err
-	}
-	return release, nil
-}
-
-func DownloadCorectlAsset(asset *CoreCtlAsset) (io.ReadCloser, error) {
-	log.Debug().Msgf("starting download %s", asset.Url)
-	resp, err := http.Get(asset.Url)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to download corectl release: %v", err)
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("failed to download corectl release: status code %v", resp.StatusCode)
-	}
-	log.Debug().Msgf("downloaded %s: %+v", asset.Url, resp)
-
-	return resp.Body, err
-}
-
-func DecompressCorectlAssetInMemory(tarData io.ReadCloser) (*tar.Reader, error) {
-	log.Debug().Msg("decompressing asset")
-
-	gzr, err := gzip.NewReader(tarData)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create gzip reader: %v", err)
-	}
-	defer gzr.Close()
-	tarReader := tar.NewReader(gzr)
-
-	for {
-		header, err := tarReader.Next()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return nil, fmt.Errorf("failed to read tar archive: %v", err)
-		}
-
-		if filepath.Base(header.Name) == "corectl" && header.Typeflag == tar.TypeReg {
-			log.Debug().Msg("found corectl in tar")
-			return tarReader, nil
-		}
-	}
-	return nil, fmt.Errorf("corectl binary not found in the release")
-}
-
-func WriteCorectlAssetToPath(tarReader *tar.Reader, tmpPath string, outFile *os.File) error {
-	binaryName := "corectl"
-
-	written, err := io.Copy(outFile, tarReader)
-	if err != nil {
-		return fmt.Errorf("failed to copy %s binary: %v", binaryName, err)
-	}
-
-	log.Debug().Msgf("%d bytes written to %s", written, tmpPath)
-
-	if err := os.Chmod(tmpPath, 0755); err != nil {
-		return fmt.Errorf("failed to set executable permissions on %s binary: %v", binaryName, err)
-	}
-
-	log.Debug().Msgf("%s has been installed to %s", binaryName, tmpPath)
-	return nil
 }
 
 func CreateGitHubPR(client *github.Client, title string, body string, branchName string, repoName string, organization string, dryRun bool) (*github.PullRequest, error) {
